chore: rename constants from default to app
